FileManagement
- extracted file management from BankApplication class
