+++ src/BankApplication.java
@@ -32,8 +32,7 @@ import net.miginfocom.swing.MigLayout;
 
 public class BankApplication extends JFrame {
 	
-	private final static int TABLE_SIZE = 29;
-	private int accouts_counter = 0;
+	private int accounts_counter = 0;
 	private int last_account_id = 0;
 	private JMenuBar menuBar;
 	private JMenu navigateMenu, recordsMenu, transactionsMenu, fileMenu, exitMenu;
@@ -67,6 +66,82 @@ public class BankApplication extends JFrame {
 		ba.setSize(1200,400);
 		ba.pack();
 		ba.setVisible(true);
+	}
+	
+	public void saveOpenValues(){		
+		if (openValues){
+			surnameTextField.setEditable(false);
+			firstNameTextField.setEditable(false);
+				
+			table.get(currentItem).setSurname(surnameTextField.getText());
+			table.get(currentItem).setFirstName(firstNameTextField.getText());
+		}
+	}	
+	
+	public void displayDetails(int key) {	
+		accountIDTextField.setText(table.get(key).getAccountID()+"");
+		accountNumberTextField.setText(table.get(key).getAccountNumber());
+		surnameTextField.setText(table.get(key).getSurname());
+		firstNameTextField.setText(table.get(key).getFirstName());
+		accountTypeTextField.setText(table.get(key).getAccountType());
+		balanceTextField.setText(table.get(key).getBalance()+"");
+		
+		if(accountTypeTextField.getText().trim().equals("Current"))
+			overdraftTextField.setText(table.get(key).getOverdraft()+"");
+		else
+			overdraftTextField.setText("Only applies to current accs");
+	}
+
+	public void writeFile(){
+		FileManagement.openFileWrite();
+		FileManagement.saveToFile(table);
+		FileManagement.closeFile();
+	}
+	
+	public void saveFileAs(){
+		FileManagement.saveToFileAs();
+		FileManagement.saveToFile(table);	
+		FileManagement.closeFile();
+	}
+	
+	public void readFile(){
+		FileManagement.openFileRead(table);
+		table = FileManagement.readRecords();
+		FileManagement.closeFile();
+		
+		accounts_counter = table.size();
+		last_account_id = table.get(accounts_counter).getAccountID();
+		
+	}
+	
+	private void searchStrategy(String searchType){
+		boolean found = false;
+		if(searchType.equals("SURNAME")){
+			String sName = JOptionPane.showInputDialog("Search for surname: ");
+			for (Map.Entry<Integer, BankAccount> entry : table.entrySet()) {
+				if(sName.equalsIgnoreCase((entry.getValue().getSurname().trim()))){
+					found = true;
+					displayDetails(entry.getKey());
+				}
+			}		
+			if(found)
+				JOptionPane.showMessageDialog(null, "Surname  " + sName + " found.");
+			else
+				JOptionPane.showMessageDialog(null, "Surname " + sName + " not found.");
+		}
+		else if(searchType.equals("ACCOUNT_NUMBER")){
+			String accNum = JOptionPane.showInputDialog("Search for account number: ");
+			for (Map.Entry<Integer, BankAccount> entry : table.entrySet()) {
+				if(accNum.equals(entry.getValue().getAccountNumber().trim())){
+					found = true;
+					displayDetails(entry.getKey());			
+				}			 
+			}
+			if(found)
+				JOptionPane.showMessageDialog(null, "Account number " + accNum + " found.");
+			else
+				JOptionPane.showMessageDialog(null, "Account number " + accNum + " not found.");
+		}
 	}
 	
 	public void initComponents() {
@@ -204,193 +279,6 @@ public class BankApplication extends JFrame {
     	setDefaultCloseOperation(EXIT_ON_CLOSE);
 	}
 	
-	public void saveOpenValues(){		
-		if (openValues){
-			surnameTextField.setEditable(false);
-			firstNameTextField.setEditable(false);
-				
-			table.get(currentItem).setSurname(surnameTextField.getText());
-			table.get(currentItem).setFirstName(firstNameTextField.getText());
-		}
-	}	
-	
-	public void displayDetails(int key) {	
-		accountIDTextField.setText(table.get(key).getAccountID()+"");
-		accountNumberTextField.setText(table.get(key).getAccountNumber());
-		surnameTextField.setText(table.get(key).getSurname());
-		firstNameTextField.setText(table.get(key).getFirstName());
-		accountTypeTextField.setText(table.get(key).getAccountType());
-		balanceTextField.setText(table.get(key).getBalance()+"");
-		
-		if(accountTypeTextField.getText().trim().equals("Current"))
-			overdraftTextField.setText(table.get(key).getOverdraft()+"");
-		else
-			overdraftTextField.setText("Only applies to current accs");
-	}
-	
-	public static void openFileRead(){
-		table.clear();
-		fc = new JFileChooser();
-		int returnVal = fc.showOpenDialog(null);
-		 
-        if (returnVal == JFileChooser.APPROVE_OPTION) {
-            File file = fc.getSelectedFile();
-        } 
-        else {
-        }
-		try{
-			if(fc.getSelectedFile()!=null)
-				input = new RandomAccessFile( fc.getSelectedFile(), "r" );
-		}
-		catch ( IOException ioException ){
-			JOptionPane.showMessageDialog(null, "File Does Not Exist.");
-		}
-	}
-	
-	public static void openFileWrite(){
-		if(fileToSaveAs!=""){
-	      try{
-	         output = new RandomAccessFile( fileToSaveAs, "rw" );
-	         JOptionPane.showMessageDialog(null, "Accounts saved to " + fileToSaveAs);
-	      }
-	      catch ( IOException ioException ){
-	    	  JOptionPane.showMessageDialog(null, "File does not exist.");
-	      }
-		}
-		else
-		   saveToFileAs();
-	}
-	
-	public static void saveToFileAs(){
-		fc = new JFileChooser();
-		
-		 int returnVal = fc.showSaveDialog(null);
-         if (returnVal == JFileChooser.APPROVE_OPTION) {
-             File file = fc.getSelectedFile();
-           
-             fileToSaveAs = file.getName();
-             JOptionPane.showMessageDialog(null, "Accounts saved to " + file.getName());
-         } 
-         else
-             JOptionPane.showMessageDialog(null, "Save cancelled by user");
-	     try {
-	    	 if(fc.getSelectedFile()==null){
-	        	JOptionPane.showMessageDialog(null, "Cancelled");
-	    	 }
-	    	 else
-	        	output = new RandomAccessFile(fc.getSelectedFile(), "rw" );
-		} 
-	    catch (FileNotFoundException e) {
-			e.printStackTrace();
-		}   
-	}
-	
-	public static void closeFile(){
-		try{
-		    if ( input != null )
-		      input.close();
-		}
-		catch ( IOException ioException ){
-			JOptionPane.showMessageDialog(null, "Error closing file.");//System.exit( 1 );
-		}
-	}
-	
-	public void readRecords(){
-		RandomAccessBankAccount record = new RandomAccessBankAccount();
-		try {
-			while ( true ){
-		        do{
-		        	if(input!=null)
-		        		record.read( input );
-		        } while ( record.getAccountID() == 0 );
-	
-	        BankAccount ba = new BankAccount(record.getAccountID(), record.getAccountNumber(), record.getFirstName(),
-	                record.getSurname(), record.getAccountType(), record.getBalance(), record.getOverdraft());
-	        
-	        accouts_counter++;
-	        last_account_id = ba.getAccountID();
-	        table.put(accouts_counter, ba);
-	     }
-	  }
-	  catch ( EOFException eofException ){
-		  return;
-	  }
-	  catch ( IOException ioException ){
-		  JOptionPane.showMessageDialog(null, "Error reading file.");
-		  System.exit( 1 );
-	  }
-   }
-	
-	public void saveToFile(){
-		RandomAccessBankAccount record = new RandomAccessBankAccount();
-	    
-		for (Map.Entry<Integer, BankAccount> entry : table.entrySet()) {
-		   record.setAccountID(entry.getValue().getAccountID());
-		   record.setAccountNumber(entry.getValue().getAccountNumber());
-		   record.setFirstName(entry.getValue().getFirstName());
-		   record.setSurname(entry.getValue().getSurname());
-		   record.setAccountType(entry.getValue().getAccountType());
-		   record.setBalance(entry.getValue().getBalance());
-		   record.setOverdraft(entry.getValue().getOverdraft());
-		   
-		   if(output!=null){
-		      try {
-					record.write( output );
-				} catch (IOException u) {
-					u.printStackTrace();
-				}
-		   }   
-		} 
-	}
-
-	public void writeFile(){
-		openFileWrite();
-		saveToFile();
-		closeFile();
-	}
-	
-	public void saveFileAs(){
-		saveToFileAs();
-		saveToFile();	
-		closeFile();
-	}
-	
-	public void readFile(){
-	    openFileRead();
-	    readRecords();
-	    closeFile();		
-	}
-	
-	private void searchStrategy(String searchType){
-		boolean found = false;
-		if(searchType.equals("SURNAME")){
-			String sName = JOptionPane.showInputDialog("Search for surname: ");
-			for (Map.Entry<Integer, BankAccount> entry : table.entrySet()) {
-				if(sName.equalsIgnoreCase((entry.getValue().getSurname().trim()))){
-					found = true;
-					displayDetails(entry.getKey());
-				}
-			}		
-			if(found)
-				JOptionPane.showMessageDialog(null, "Surname  " + sName + " found.");
-			else
-				JOptionPane.showMessageDialog(null, "Surname " + sName + " not found.");
-		}
-		else if(searchType.equals("ACCOUNT_NUMBER")){
-			String accNum = JOptionPane.showInputDialog("Search for account number: ");
-			for (Map.Entry<Integer, BankAccount> entry : table.entrySet()) {
-				if(accNum.equals(entry.getValue().getAccountNumber().trim())){
-					found = true;
-					displayDetails(entry.getKey());			
-				}			 
-			}
-			if(found)
-				JOptionPane.showMessageDialog(null, "Account number " + accNum + " found.");
-			else
-				JOptionPane.showMessageDialog(null, "Account number " + accNum + " not found.");
-		}
-	}
-	
 	private void setListeners(){
 		setOverdraft.addActionListener(setOverdraftListener);
 		nextItemButton.addActionListener(nextItemButtonActionListener);
@@ -497,9 +385,9 @@ public class BankApplication extends JFrame {
 		
 		createItemListener = new ActionListener(){
 			public void actionPerformed(ActionEvent e){
-				accouts_counter++;
+				accounts_counter++;
 				last_account_id++;
-				new CreateBankDialog(table, last_account_id, accouts_counter);
+				new CreateBankDialog(table, last_account_id, accounts_counter);
 			}
 		};
 		
@@ -669,72 +557,3 @@ public class BankApplication extends JFrame {
 		};		
 	}
 }
-
-
-//deleteItemListener, createItemListener, modifyItemListener, 
-//setInterestListener, listAllListener, openFileListener, saveFileListener, saveAsFileListener, closeAppListener, findBySurnameListener, findByAccountListener, depositAmountListener, withdrawAmountListener, calcInterestListener;
-
-
-/*
-The task for your second assignment is to construct a system that will allow users to define a data structure representing a collection of records that can be displayed both by means of a dialog that can be scrolled through and by means of a table to give an overall view of the collection contents. 
-The user should be able to carry out tasks such as adding records to the collection, modifying the contents of records, and deleting records from the collection, as well as being able to save and retrieve the contents of the collection to and from external random access files.
-The records in the collection will represent bank account records with the following fields:
-
-AccountID (this will be an integer unique to a particular account and 
-will be automatically generated when a new account record is created)
-
-AccountNumber (this will be a string of eight digits and should also 
-be unique - you will need to check for this when creating a new record)
-
-Surname (this will be a string of length 20)
-
-FirstName (this will be a string of length 20)
-
-AccountType (this will have two possible options - "Current " and 
-"Deposit" - and again will be selected from a drop down list when 
-entering a record)
-
-Balance (this will a real number which will be initialised to 0.0 
-and can be increased or decreased by means of transactions)
-
-Overdraft (this will be a real number which will be initialised 
-to 0.0 but can be updated by means of a dialog - it only applies 
-to current accounts)
-
-You may consider whether you might need more than one class to deal with bank accounts.
-The system should be menu-driven, with the following menu options:
-
-Navigate: First, Last, Next, Previous, Find By Account Number 
-(allows you to find a record by account number entered via a 
-dialog box), Find By Surname (allows you to find a record by 
-surname entered via a dialog box),List All (displays the 
-contents of the collection as a dialog containing a JTable)
-
-Records: Create, Modify, Delete, Set Overdraft (this should 
-use a dialog to allow you to set or update the overdraft for 
-a current account), Set Interest Rate (this should allow you 
-to set the interest rate for deposit accounts by means of a 
-dialog)
-
-Transactions: Deposit, Withdraw (these should use dialogs which
-allow you to specify an account number and the amount to withdraw
-or deposit, and should check that a withdrawal would not cause
-the overdraft limit for a current account to be exceeded, or be 
-greater than the balance in a deposit account, before the balance 
-is updated), Calculate Interest (this calculates the interest rate 
-for all deposit accounts and updates the balances)
-
-File: Open, Save, Save As (these should use JFileChooser dialogs. 
-The random access file should be able to hold 25 records. The position 
-in which a record is stored and retrieved will be determined by its account 
-number by means of a hashing procedure, with a standard method being used when 
-dealing with possible hash collisions)
-
-Exit Application (this should make sure that the collection is saved - or that 
-the user is given the opportunity to save the collection - before the application closes)
-
-When presenting the results in a JTable for the List All option, the records should be sortable on all fields, but not editable (changing the records or adding and deleting records can only be done through the main dialog).
-For all menu options in your program, you should perform whatever validation, error-checking and exception-handling you consider to be necessary.
-The programs Person.java and PersonApplication.java (from OOSD2) and TableDemo.java may be of use to you in constructing your interfaces. The set of Java programs used to create, edit and modify random access files will also provide you with a basis for your submission.
-
-*/
